Refactored routes, requests now don't need to pass the user id anymore

+++ src/main/kotlin/com/enestekin/routes/PostsRoutes.kt
@@ -4,7 +4,6 @@ package com.enestekin.routes
 import com.enestekin.data.requests.CreatePostRequest
 import com.enestekin.data.requests.DeletePostRequest
 import com.enestekin.data.responses.BasicApiResponse
-import com.enestekin.plugins.email
 import com.enestekin.service.LikeService
 import com.enestekin.service.UserService
 import com.enestekin.util.ApiResponseMessages
@@ -20,11 +19,10 @@ import io.ktor.routing.*
 
 fun Route.createPost(
     postService: PostService,
-    userService: UserService
-){
+) {
 
     authenticate {
-        post("/api/post/create"){
+        post("/api/post/create") {
             val request = call.receiveOrNull<CreatePostRequest>() ?: kotlin.run {
                 call.respond(HttpStatusCode.BadRequest)
                 return@post
@@ -36,90 +34,63 @@ when we attached a claim  which is email.user logs in . email saves in token.
  *
  **/
 
-                // same *
-//            val email = call.principal<JWTPrincipal>()?.getClaim("email",String::class)
-//            val isEmailByUser =userService.doesEmailBelongToUserId(
-//                email =  email ?: "",
-//                userId = request.userId
-//            )
-//
-//            if (!isEmailByUser){
-//                call.respond(HttpStatusCode.Unauthorized,"You are not who you say you are.")
-//                return@post
-//            }
 
-                // same *
-            println("${request.userId} %%%%% ${request.description}")
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ){
-                val didUserExist = postService.createPostIfUserExists(request)
-                println("$didUserExist %%%%% .........")
+            val userId = call.userId
 
-                if (!didUserExist){
-                    println("$didUserExist false")
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(
-                            successful = false,
-                            message = ApiResponseMessages.USER_NOT_FOUND
-                        )
+            val didUserExist = postService.createPostIfUserExists(request,userId)
+
+            if (!didUserExist) {
+                println("$didUserExist false")
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(
+                        successful = false,
+                        message = ApiResponseMessages.USER_NOT_FOUND
                     )
-                }else {
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(
-                            successful = true,
-                        )
+                )
+            } else {
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(
+                        successful = true,
                     )
-                }
+                )
             }
-
         }
-    }
 
+    }
 }
+
+
 fun Route.getPostsForFollows(
     postService: PostService,
-    userService: UserService
-){
-   authenticate {
-       get ("/api/post/get"){
-           val userId = call.parameters[QueryParams.PARAM_USER_ID] ?: kotlin.run {
-               call.respond(HttpStatusCode.BadRequest)
-               return@get
-           }
-           val page = call.parameters[QueryParams.PARAM_PAGE]?.toIntOrNull() ?: 0
-           val pageSize = call.parameters[QueryParams.PARAM_PAGE_SIZE]?.toIntOrNull() ?:
-           Constants.DEFAULT_POST_PAGE_SIZE
+) {
+    authenticate {
+        get("/api/post/get") {
 
 
+            val page = call.parameters[QueryParams.PARAM_PAGE]?.toIntOrNull() ?: 0
+            val pageSize =
+                call.parameters[QueryParams.PARAM_PAGE_SIZE]?.toIntOrNull() ?: Constants.DEFAULT_POST_PAGE_SIZE
 
-//ifEmailBelongsToUser  is basically validating  email shorter.
-           ifEmailBelongsToUser(
-               userId = userId,
-               validateEmail = userService::doesEmailBelongToUserId
-           ){
-               println("Calisti")
-               val posts = postService.getPostsForFollows(userId,page,pageSize)
-               call.respond(
-                   HttpStatusCode.OK,
-                   posts
-               )
-           }
 
-       }
-   }
+            val posts = postService.getPostsForFollows(call.userId, page, pageSize)
+            call.respond(
+                HttpStatusCode.OK,
+                posts
+            )
+        }
+
+    }
 }
+
 
 fun Route.deletePost(
     postService: PostService,
-    userService: UserService,
     likeService: LikeService
-){
-
-    delete ("/api/post/delete") {
+) {
+    authenticate {
+    delete("/api/post/delete") {
         val request = call.receiveOrNull<DeletePostRequest>() ?: kotlin.run {
             call.respond(HttpStatusCode.BadRequest)
             return@delete
@@ -127,20 +98,26 @@ fun Route.deletePost(
         }
 
         val post = postService.getPost(request.postId)
+
+
         if (post == null) {
             call.respond(
                 HttpStatusCode.NotFound
             )
             return@delete
         }
-        ifEmailBelongsToUser(
-            userId = post.userId,
-            validateEmail = userService::doesEmailBelongToUserId
-        ) {
+        if (post.userId == call.userId){
+
             postService.deletePost(request.postId)
             likeService.deleteLikesForParent(request.postId)
             // TODO: 28.10.2021 Delete comments from post
             call.respond(HttpStatusCode.OK)
+        }else {
+            call.respond(HttpStatusCode.Unauthorized)
         }
+
     }
+    }
+
+
 }
